feat: directly inject custom markdown with expandable content
- Replace modal-based custom markdown with direct injection
- Add new ExpandableContent UI component for max-height + expand/collapse behavior
- Extract CollapsedSection from RevisionDiffUtils into reusable ExpandableContent
- Update CustomMarkdown to use ExpandableContent with configurable maxHeight
- The markdown is now visible inline on the page with a fade-out gradient
  and 'Show more/Show less' toggle when content overflows
