+++ packages/front-end/ui/ExpandableContent.module.scss
@@ -1,0 +1,8 @@
+.fadeOverlay {
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  height: 64px;
+  pointer-events: none;
+}
